+++ pg_yregress/CHANGELOG.md
@@ -8,6 +8,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+* Support for unmanaged (external) instances
+
 ## [0.1.0] - 2023-08-25
 
 Initial release following a few months of iterative development.